impl: isola api para futura transf em interlegis/drfautoapi
