--- sapl/api/core/schema.py
@@ -1,5 +1,0 @@
-from drf_spectacular.openapi import AutoSchema
-
-
-class Schema(AutoSchema):
-    pass
